When loading data into the DSS, best results are obtained when all the cloud URLs loaded by reference are accessible. This provides the most complete set of file metadata for the files loaded by reference. However, there are times, mainly during development and testing, when that is not the case. Currently, as we load the GTEx data, we have access to the NIH files in Google but not in AWS. The most important file information is included in the loader input data already. This change enables the loader to run and do a generally adequate job even when not all the cloud URLs are accessible.
The specific changes are:

Make use of the data file size from the input data. File size was already a required field
for the input schema, but was not be used. Now the input file size is compared to the
actual file size of the cloud URLs that are accessible, and discrepancies result in an
error loading the bundle.
If no actual cloud URLs are available for the a given data file, the input file size is used.

Improved warnings when the cloud URLs are not accessible, using the Python warnings feature/library.

Updated and enhanced unit tests for the above