Correct/improve support for the `--dry-run` option (resolves #46)
The loader --dry-run option is very useful for checking that the many
preconditions for a successful load before actually putting files/bundles into the DSS.
This is especially important because the DSS doesn't currently provide a good way
to clean-up a partial or unsuccessful load.

However, the implementation of --dry-run was in disrepair and needed to be updated/repaired.
Multiple small yet important changes were made, and a unit test was added.